Adds an interface for the vulnerable items.
The vulnerable items can reiceive an attack on one of their sides.

ptb::monster is vulnerable. If the attacker is a monster, it processes the
attack as before. Otherwise it does nothing.

ptb::monster_item attacks the vulnerables.

ptb::rabbit is vulnerable, thus it can be attacked by a monster (read: the
player or a stone) and does not have to handle the collision itself.